Add files via upload
This is a program that allows the user to input names, and then randomly selects one of the names to go first. This program notes the order that the names were typed into the prompt, and will continue cycling through the names in that same order, but will start with one of those names randomly.